pyproject.toml
[project]
name = "cuda-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.136.0",
    "loguru>=0.7.3",
    "nvidia-nat[langchain]>=1.6.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
